cleanup code and add comments to blog views

+++ blog/views.py
@@ -10,19 +10,25 @@ from django.http import HttpResponse, HttpResponseRedirect
 def blog(request):
     """ View for displaying the blog page """
     posts = BlogPost.objects.all()
+    # Get all posts to show when rendering the template
     form = BlogForm(request.POST or None)
+    # Access the blog form
     context = {'form': form, 'posts': posts}
+    # Context for rendering template
     if request.method == 'POST':
         if request.user.is_superuser:
             title = request.POST.get('title')
             image = request.FILES.get('image')
             body = request.POST.get('body')
+            # Access the title, image and body from the form
             blog = BlogPost.objects.create(
                 title=title,
                 image=image,
                 body=body,
             )
+            # Create a new blog post object with the data from the form
             blog.save()
+            # Save blog post
             messages.success(request, 'Blog successfully posted!')
             return HttpResponseRedirect('')
         else:
@@ -33,23 +39,29 @@ def blog(request):
 
 @login_required
 def remove_blog_post(request, item_id):
-    """Remove the item from the shopping bag"""
+    """Remove blog post"""
     posts = BlogPost.objects.all()
+    # Get all posts to show when rendering the template
     form = BlogForm(request.POST or None)
+    # Access the blog form
     context = {'form': form, 'posts': posts}
-    if request.method == 'POST':
-        if request.user.is_superuser:
-            try:
-                post = get_object_or_404(BlogPost, pk=item_id)
-                image = post.image
-                messages.success(
-                    request, f'Removed {post.title}')
-                image.delete()
-                post.delete()
-                return HttpResponseRedirect('')
-            except Exception as e:
-                messages.error(request, f'Error removing post: {e}')
-                return HttpResponse(status=500)
+    # Context for rendering template
+    if request.method == 'POST' and request.user.is_superuser:
+        try:
+            post = get_object_or_404(BlogPost, pk=item_id)
+            # Access the blog post related to the id in the template
+            image = post.image
+            # Access the image for the blog post
+            image.delete()
+            # Delete the image
+            post.delete()
+            # Delete the post
+            messages.success(
+                request, f'Removed {post.title}')
+            return HttpResponseRedirect('')
+        except Exception as e:
+            messages.error(request, f'Error removing post: {e}')
+            return HttpResponse(status=500)
     return render(request, 'blog/blog.html', context)
 
 
@@ -57,12 +69,16 @@ def remove_blog_post(request, item_id):
 def edit_blog_post(request, item_id):
     """ Edit blog post """
     post = get_object_or_404(BlogPost, pk=item_id)
+    # Get the current blog post
     form = BlogForm(instance=post)
+    # Populate the form with the data from the post instance
     if request.method == 'POST' and request.user.is_superuser:
         form = BlogForm(request.POST, request.FILES, instance=post)
+        # Get the new form data
         if form.has_changed and form.is_valid():
             try:
                 post = form.save()
+                # Save the new form data to the original post
                 messages.success(
                     request, f'Successfully edited {post.title}')
                 return redirect(reverse('blog'))
